+++ server/api/users.js
@@ -1,5 +1,6 @@
 const router = require('express').Router()
 const {User} = require('../db/models')
+const isAdmin = require('./adminCheck')
 module.exports = router
 
 router.get('/', async (req, res, next) => {
