+++ server/auth/google.js
@@ -32,14 +32,29 @@ if (!process.env.GOOGLE_CLIENT_ID || !process.env.GOOGLE_CLIENT_SECRET) {
     (token, refreshToken, profile, done) => {
       const googleId = profile.id
       const email = profile.emails[0].value
-      const imgUrl = profile.photos[0].value
+      // let imgUrl = profile.photos[0].value //not needed
       const firstName = profile.name.givenName
       const lastName = profile.name.familyName
-      const fullName = profile.displayName
+      // let fullName = profile.displayName //not needed
+      let addressLine1 = '(empty)'
+      let city = '(empty)'
+      let state = '(empty)'
+      let zip = '(empty)'
+      // let phone = '(empty)'
+      let phone = profile.phone
 
       User.findOrCreate({
         where: {googleId},
-        defaults: {email, imgUrl, firstName, lastName, fullName}
+        defaults: {
+          email,
+          firstName,
+          lastName,
+          addressLine1,
+          city,
+          state,
+          zip,
+          phone
+        }
       })
         .then(([user]) => done(null, user))
         .catch(done)
