+++ server/index.js
@@ -10,6 +10,7 @@ const sessionStore = new SequelizeStore({db})
 const PORT = process.env.PORT || 8080
 const app = express()
 const socketio = require('socket.io')
+const User = require('./db/models/user')
 module.exports = app
 
 // This is a global Mocha hook, used for resource cleanup.
@@ -33,7 +34,8 @@ passport.serializeUser((user, done) => done(null, user.id))
 
 passport.deserializeUser(async (id, done) => {
   try {
-    const user = await db.models.user.findByPk(id)
+    // const user = await db.models.user.findByPk(id)
+    const user = await User.findByPk(id)
     done(null, user)
   } catch (err) {
     done(err)
